requirements.txt
boto3
mysql-connector
PyQt5
python-dotenv
